Use a single buffer pair for lines instead of one for every instance. Changed the scene to a vector rather than a sorted multiset (too slow).

+++ VK/Texture.hh
@@ -13,9 +13,9 @@ namespace Animate::VK
 
             void load_from_resource(std::string resource_id);
             void use();
-            GLuint get_id();
+            uint64_t get_id();
 
         protected:
-            GLuint texture_id;
+            uint64_t texture_id;
     };
 }
